+++ contracts/community-sbt/CHANGELOG.md
@@ -23,6 +23,7 @@ Change log entries are to be added to the Unreleased section. Example entry:
 
 ### Breaking Changes
 
+- Contract field type for `admin` changed from `AccoundId` -> `LazyOption<Vec<AccountId>>`
 - Recommended `cost.mint_deposit` is decreased by 0.001 miliNEAR (in total).
 
 ### Bug Fixes